Refactor Responses API support
